+++ src/js/modwatch.module.js
@@ -1,7 +1,7 @@
 import AppConfig from "./config/appConfig";
 
-import "angular";
-import "angular-route";
+// import "angular";
+// import "angular-route";
 // import "angular-animate";
 
 import "angular-local-storage";